+++ src/main/java/com/codenear/butterfly/member/domain/repository/member/MemberRepository.java
@@ -11,6 +11,8 @@ import java.util.Optional;
 public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
     Optional<Member> findByNickname(String nickname);
     Optional<Member> findByEmailAndPlatform(String email, Platform platform);
+    Optional<Member> findByPhoneNumberAndPlatform(String phoneNumber, Platform platform);
     Optional<Member> findByPhoneNumber(String phoneNumber);
     boolean existsByEmail(String email);
+    boolean existsByPhoneNumber(String phoneNumber);
 }
